% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_err}
\alias{compute_err}
\title{Compute the items suggested to be removed / revised based on item infit / pvalues}
\usage{
compute_err(sim_res, cutoff = c(0.8, 1.2))
}
\arguments{
\item{sim_res:}{simulation results}

\item{cutoff:}{itemfit cutoff}
}
\value{
returns a list with alpha / beta errors across all
and the number of items removed (for each matrix)
}
\description{
Compute the items suggested to be removed / revised based on item infit / pvalues
}
\examples{
compute_err(sim_res)
}

