% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{estimate}
\alias{estimate}
\title{wrapper around RM function from eRm}
\usage{
estimate(n, items, model_sim, ...)
}
\arguments{
\item{n,}{items: numeric}

\item{model_sim:}{function to simulate Rasch data}

\item{...:}{additional arguments to model_sim}
}
\value{
returns the estimated eRm object
}
\description{
tryCatch until the simulated data matrix is neither ill-conditioned
nor has a participant with all 0 or all 1
}
\examples{
estimate(400, 30, sim.2pl, .5)
}

