% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{get_itemfit}
\alias{get_itemfit}
\title{Get the itemfit statistics (infit, outfit, p-values)}
\usage{
get_itemfit(n, items, times, parallel, model_sim, ...)
}
\arguments{
\item{n,}{items, times: numeric}

\item{model_sim:}{function to simulate Rasch data}

\item{...:}{additional arguments to model_sim}
}
\value{
returns an object of class 'sim_res'
which is an array that stores #times #items x 3 matrices
}
\description{
Get the itemfit statistics (infit, outfit, p-values)
}
\examples{
get_itemfit(400, 30, sim.2pl, .4)
}

