% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{visualize}
\alias{visualize}
\title{Visualizes the \% items removed / Rasch rejected given a simulation result}
\usage{
visualize(res_sum, ylab = "\% removed items", ...)
}
\arguments{
\item{res_sum:}{summarized simulation result or processed LR simulation}

\item{ylab:}{label of the y-axis}

\item{...:}{additional arguments to xyplot}
}
\value{
plots a trellis graph visualising the results
}
\description{
Visualizes the \% items removed / Rasch rejected given a simulation result
}
\examples{
visualize(res_sum)
}

